Remove dataframe protocol abstract classes and move the docstrings and necessary defenitions to separate implementation files
